fan -r now uploads the report #89
