--- FanControl/Features/Report/Models/AppReportRequestBody.swift
@@ -1,4 +1,0 @@
-struct AppReportRequestBody: Encodable {
-    let hash: String
-    let report: String
-}
